Add android-x86 plugins (www.android-x86.org).
android-fromiso will detect the android-x86 ISO filenames and add them for you, but it only works with 4.4 and up. android-fromroot can be used with lower versions, but you can only use it once per disc (it needs to put install.img, ramdisk.img, and system.sfs in the root.) To use android-fromroot, call your ISO "android-fromroot.iso" instead of the default name.